backend/requirementsv8_3.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
stripe==11.4.1
cryptography==44.0.0
python-multipart==0.0.18
aiofiles==24.1.0
psycopg2-binary==2.9.10
slowapi==0.1.9
